requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-multipart==0.0.9
openai==1.82.0
python-docx==1.1.0
PyPDF2==3.0.1
psycopg2-binary==2.9.10
sqlalchemy==2.0.35
pydantic==2.9.0
python-dotenv==1.0.1
pytest==7.4.3